sql: refactor regionutils into separate pkg for new schema changer
Previously, the sql/regionutils.go contained all the logic
required to help setup and create zone configs. These changes
split off that code into its own package, so that the declarative
schema changer can access these functions without cyclic dependencies.

Release note: None
