+++ pkg/sql/BUILD.bazel
@@ -326,7 +326,6 @@ go_library(
         "//pkg/sql/contention",
         "//pkg/sql/contention/txnidcache",
         "//pkg/sql/contentionpb",
-        "//pkg/sql/covering",
         "//pkg/sql/decodeusername",
         "//pkg/sql/delegate",
         "//pkg/sql/descmetadata",
@@ -371,6 +370,7 @@ go_library(
         "//pkg/sql/physicalplan/replicaoracle",
         "//pkg/sql/privilege",
         "//pkg/sql/querycache",
+        "//pkg/sql/regionutils",
         "//pkg/sql/roleoption",
         "//pkg/sql/row",
         "//pkg/sql/rowcontainer",
